Refactor structs, fix clippy
